+++ src/main/resources/assets/js/app/inputtype/text/HtmlArea.ts
@@ -25,18 +25,12 @@ import {HtmlEditor} from '../ui/text/HtmlEditor';
 import {HtmlEditorParams} from '../ui/text/HtmlEditorParams';
 import {StylesRequest} from '../ui/text/styles/StylesRequest';
 import {BaseInputTypeNotManagingAdd} from 'lib-admin-ui/form/inputtype/support/BaseInputTypeNotManagingAdd';
-import {LoginResult} from 'lib-admin-ui/security/auth/LoginResult';
-import {IsAuthenticatedRequest} from 'lib-admin-ui/security/auth/IsAuthenticatedRequest';
 import {TextArea} from 'lib-admin-ui/ui/text/TextArea';
 import {FormInputEl} from 'lib-admin-ui/dom/FormInputEl';
 import {SelectorOnBlurEvent} from 'lib-admin-ui/ui/selector/SelectorOnBlurEvent';
 import {BrowserHelper} from 'lib-admin-ui/BrowserHelper';
 import {FormEl} from 'lib-admin-ui/dom/FormEl';
 import {ArrayHelper} from 'lib-admin-ui/util/ArrayHelper';
-import {ProjectGetRequest} from '../../settings/resource/ProjectGetRequest';
-import {ProjectContext} from '../../project/ProjectContext';
-import {Project} from '../../settings/data/project/Project';
-import {PrincipalKey} from 'lib-admin-ui/security/PrincipalKey';
 
 declare var CONFIG;
 
@@ -67,37 +61,12 @@ export class HtmlArea
 
         this.inputConfig = config.inputConfig;
 
-        this.authRequest =
-            Q.all([new IsAuthenticatedRequest().sendAndParse(), new ProjectGetRequest(ProjectContext.get().getProject()).sendAndParse()])
-                .spread((loginResult: LoginResult, project: Project) => {
-                    this.editableSourceCode = this.isAllowedToEditSourceCode(loginResult, project);
-
-                    return null;
-                });
+        this.authRequest = HTMLAreaHelper.isSourceCodeEditable().then((value: boolean) => {
+            this.editableSourceCode = value;
+            return Q(null);
+        });
 
         this.setupEventListeners();
-    }
-
-    private isAllowedToEditSourceCode(loginResult: LoginResult, project: Project): boolean {
-        if (loginResult.isContentExpert()) {
-            return true;
-        }
-
-        const userPrincipals: PrincipalKey[] = loginResult.getPrincipals();
-
-        const isExpert: boolean = project.getPermissions().getExperts().some((expert: PrincipalKey) => {
-            return userPrincipals.some((userPrincipal: PrincipalKey) => userPrincipal.equals(expert));
-        });
-
-        if (isExpert) {
-            return true;
-        }
-
-        const isOwner: boolean = project.getPermissions().getOwners().some((owner: PrincipalKey) => {
-            return userPrincipals.some((userPrincipal: PrincipalKey) => userPrincipal.equals(owner));
-        });
-
-        return isOwner;
     }
 
     private setupEventListeners() {
