+++ src/main/resources/assets/js/app/inputtype/ui/text/HTMLAreaHelper.ts
@@ -4,6 +4,14 @@ import {ContentId} from 'lib-admin-ui/content/ContentId';
 import {ImageUrlResolver} from '../../../util/ImageUrlResolver';
 import {Styles} from './styles/Styles';
 import {UriHelper} from 'lib-admin-ui/util/UriHelper';
+import * as Q from 'q';
+import {ProjectGetRequest} from '../../../settings/resource/ProjectGetRequest';
+import {ProjectContext} from '../../../project/ProjectContext';
+import {Project} from '../../../settings/data/project/Project';
+import {DefaultErrorHandler} from 'lib-admin-ui/DefaultErrorHandler';
+import {IsAuthenticatedRequest} from 'lib-admin-ui/security/auth/IsAuthenticatedRequest';
+import {LoginResult} from 'lib-admin-ui/security/auth/LoginResult';
+import {PrincipalKey} from 'lib-admin-ui/security/PrincipalKey';
 
 export class HTMLAreaHelper {
 
@@ -96,5 +104,38 @@ export class HTMLAreaHelper {
         });
 
         return processedContent;
+    }
+
+    public static isSourceCodeEditable(): Q.Promise<boolean> {
+        return new IsAuthenticatedRequest().sendAndParse().then((loginResult: LoginResult) => {
+            if (loginResult.isContentExpert()) {
+                return Q(true);
+            }
+
+            return new ProjectGetRequest(ProjectContext.get().getProject()).sendAndParse().then((project: Project) => {
+                return Q(HTMLAreaHelper.isAllowedToEditSourceCode(loginResult, project));
+            });
+        }).catch((reason: any) => {
+            DefaultErrorHandler.handle(reason);
+            return Q(false);
+        });
+    }
+
+    private static isAllowedToEditSourceCode(loginResult: LoginResult, project: Project): boolean {
+        const userPrincipals: PrincipalKey[] = loginResult.getPrincipals();
+
+        const isExpert: boolean = project.getPermissions().getEditors().some((expert: PrincipalKey) => {
+            return userPrincipals.some((userPrincipal: PrincipalKey) => userPrincipal.equals(expert));
+        });
+
+        if (isExpert) {
+            return true;
+        }
+
+        const isOwner: boolean = project.getPermissions().getOwners().some((owner: PrincipalKey) => {
+            return userPrincipals.some((userPrincipal: PrincipalKey) => userPrincipal.equals(owner));
+        });
+
+        return isOwner;
     }
 }
